Started Pause menu/Reorganize
Reorganized where some stuff was so it is easier to navigate. added a scripts folder where we can put all our scripts and farther organize it with sub folders.
Added cinemachine which makes it a lot easier to do camera controls and add things like event/cutscenes.
started making a pause menu, added another action map in the input system that will control stuff when the game is paused so the player isnt effected.