Working Doors and Beginning of Pistachios
I have the doors able to open from the press of a button, which I am working to get it working with a shot to the button instead. I have the pistachios able to be collected and added to the amount of available ammo, and I'm currently making it so that the gun can actually instantiate bullets so that I can get the events and ammo working properly with visual cues.