Player Health
I made the scene workable. Everything else I think is an issue with the old shotgun, but all the events are working. Players now gain health when they pick up a pistachio. Since this is all done with older stuff, we will have to double-check how things work when adding the new scripts/changes