Better titlescreen and fixed pause menu opening
Created script that made rats spawn in title screen.
chooses their position and speed according to current screen size with a base value that it uses that can be easily changed.
puts the rats behind the signs based off of the y-axis value they are spawned at.
extended pause screen rope and made it come into view before pause menu itself to look better.
set the canvas that the player UI is to false and active before pause menu spawns and after it moves off screen to look nicer.